add tests for school user search
